Fix Deployment in Garden Runtime Cluster (#767)
* Fix `StorageClass` deployment

The `StorageClass` must not be considered if deployed
by the `gardener-operator` to the runtime cluster.
Otherwise different extensions fight against creating
the same resource in the cluster.

* Drop default values for seed and runtime cluster

The default values shouldn't anticipate if the extension
is deployed to a seed or garden runtime cluster.
Dropping the default values makes this chart handling
less error-prone.

* [make generate]